# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Knowora Universal AI Master Manager",
    "version": "3.0.0",
    "description": "Complete AI Action API to Read all blogs, Detect Errors, Update/Edit posts, Delete unwanted posts, and Publish new verified articles on Knowora.in."
  },
  "servers": [
    {
      "url": "https://www.knowora.in"
    }
  ],
  "paths": {
    "/api/gpt/blogs": {
      "get": {
        "operationId": "listAllBlogs",
        "summary": "Read and analyze all blogs for quality and errors",
        "description": "Retrieves blogs with word count analysis, detected issues (thin content, missing tables, missing FAQs, broken links), and status.",
        "parameters": [
          {
            "name": "secret",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "default": "knowora-secret-2026"
            },
            "description": "API Secret Key"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 30
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "filterQuality",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "all",
                "errors_only",
                "thin_only"
              ]
            },
            "description": "Filter to view only blogs with errors or thin content"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Search keyword"
          }
        ],
        "responses": {
          "200": {
            "description": "List of blogs with quality analysis"
          }
        }
      }
    },
    "/api/gpt/blogs/{id}": {
      "get": {
        "operationId": "getBlogDetails",
        "summary": "Read full content of a specific blog",
        "description": "Returns the complete HTML body, metadata, and fields of a blog post by ID or slug.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Blog ID or URL slug"
          },
          {
            "name": "secret",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "default": "knowora-secret-2026"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Full blog content"
          }
        }
      },
      "put": {
        "operationId": "updateBlog",
        "summary": "Edit and fix errors in an existing blog",
        "description": "Update the title, content HTML, excerpt, SEO metadata, or status of an existing blog to fix mistakes and expand content.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Blog ID or Slug to edit"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "secret"
                ],
                "properties": {
                  "secret": {
                    "type": "string",
                    "default": "knowora-secret-2026"
                  },
                  "title": {
                    "type": "string",
                    "description": "Updated Hindi SEO title"
                  },
                  "content": {
                    "type": "string",
                    "description": "Complete updated HTML content with tables and FAQ"
                  },
                  "excerpt": {
                    "type": "string",
                    "description": "Updated summary"
                  },
                  "gridBox": {
                    "type": "string",
                    "enum": [
                      "latestJobs",
                      "admitCard",
                      "examResults",
                      "scholarship",
                      "scheme",
                      "tech",
                      "finance"
                    ]
                  },
                  "seoTitle": {
                    "type": "string"
                  },
                  "seoDescription": {
                    "type": "string"
                  },
                  "seoKeywords": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "Published",
                      "Draft"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Blog updated successfully"
          }
        }
      },
      "delete": {
        "operationId": "deleteBlog",
        "summary": "Delete an unwanted or poor quality blog",
        "description": "Permanently removes a blog post by ID or slug.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "secret",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "default": "knowora-secret-2026"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Blog deleted successfully"
          }
        }
      }
    },
    "/api/publish-from-ai": {
      "post": {
        "operationId": "publishNewBlog",
        "summary": "Publish a new comprehensive blog post",
        "description": "Publishes a 2000+ words Hindi master blog with tables, FAQs, and official links.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "secret",
                  "title",
                  "content"
                ],
                "properties": {
                  "secret": {
                    "type": "string",
                    "default": "knowora-secret-2026"
                  },
                  "title": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "excerpt": {
                    "type": "string"
                  },
                  "gridBox": {
                    "type": "string"
                  },
                  "slug": {
                    "type": "string"
                  },
                  "metaTitle": {
                    "type": "string"
                  },
                  "metaDescription": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Blog published successfully"
          }
        }
      }
    }
  }
}